+++ jgeohash-core/src/main/java/de/alpharogroup/jgeohash/model/FirstAndSecondRingRegion.java
@@ -22,8 +22,8 @@ import lombok.Getter;
 import lombok.ToString;
 
 /**
- * The class {@link FirstAndSecondRingRegion} provides the neighbors from the given geohash value in first
- * and second ring.
+ * The class {@link FirstAndSecondRingRegion} provides the neighbors from the given geohash value in
+ * first and second ring.
  */
 @Getter
 @EqualsAndHashCode(callSuper = true)