+++ jgeohash-core/src/test/java/de/alpharogroup/jgeohash/model/FirstRingRegionTest.java
@@ -19,12 +19,13 @@ import static org.testng.AssertJUnit.assertEquals;
 
 import org.testng.annotations.Test;
 
+import de.alpharogroup.AbstractTestCase;
 import de.alpharogroup.evaluate.object.EqualsHashCodeAndToStringEvaluator;
 
 /**
  * The unit test class for the class {@link FirstRingRegion}
  */
-public class FirstRingRegionTest
+public class FirstRingRegionTest extends AbstractTestCase<String, String>
 {
 
 	/**
@@ -46,4 +47,174 @@ public class FirstRingRegionTest
 		expected = true;
 		assertEquals(expected, actual);
 	}
+
+	/**
+	 * Test method for {@link FirstRingRegion#getCenter()}.
+	 */
+	@Test
+	public void testGetCenter()
+	{
+		FirstRingRegion geoHashRegion = new FirstRingRegion("u");
+
+		expected = "u";
+		actual = geoHashRegion.getCenter();
+		assertEquals(expected, actual);
+
+		geoHashRegion = new FirstRingRegion("t");
+
+		expected = "t";
+		actual = geoHashRegion.getCenter();
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link FirstRingRegion#getEast()}.
+	 */
+	@Test
+	public void testGetEast()
+	{
+		FirstRingRegion geoHashRegion = new FirstRingRegion("u");
+
+		expected = "v";
+		actual = geoHashRegion.getEast();
+		assertEquals(expected, actual);
+
+		geoHashRegion = new FirstRingRegion("t");
+
+		expected = "w";
+		actual = geoHashRegion.getEast();
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link FirstRingRegion#getNorth()}.
+	 */
+	@Test
+	public void testGetNorth()
+	{
+		FirstRingRegion geoHashRegion = new FirstRingRegion("u");
+		expected = "h";
+		actual = geoHashRegion.getNorth();
+		assertEquals(expected, actual);
+
+		geoHashRegion = new FirstRingRegion("t");
+		expected = "v";
+		actual = geoHashRegion.getNorth();
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link FirstRingRegion#getNorthEast()}.
+	 */
+	@Test
+	public void testGetNorthEast()
+	{
+		FirstRingRegion geoHashRegion = new FirstRingRegion("u");
+
+		expected = "j";
+		actual = geoHashRegion.getNorthEast();
+		assertEquals(expected, actual);
+
+		geoHashRegion = new FirstRingRegion("t");
+
+		expected = "y";
+		actual = geoHashRegion.getNorthEast();
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link FirstRingRegion#getNorthWest()}.
+	 */
+	@Test
+	public void testGetNorthWest()
+	{
+		FirstRingRegion geoHashRegion = new FirstRingRegion("u");
+
+		expected = "5";
+		actual = geoHashRegion.getNorthWest();
+		assertEquals(expected, actual);
+
+		geoHashRegion = new FirstRingRegion("t");
+
+		expected = "u";
+		actual = geoHashRegion.getNorthWest();
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link FirstRingRegion#getSouth()}.
+	 */
+	@Test
+	public void testGetSouth()
+	{
+		FirstRingRegion geoHashRegion = new FirstRingRegion("u");
+
+		expected = "s";
+		actual = geoHashRegion.getSouth();
+		assertEquals(expected, actual);
+
+		geoHashRegion = new FirstRingRegion("t");
+
+		expected = "m";
+		actual = geoHashRegion.getSouth();
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link FirstRingRegion#getSouthEast()}.
+	 */
+	@Test
+	public void testGetSouthEast()
+	{
+		FirstRingRegion geoHashRegion = new FirstRingRegion("u");
+
+		expected = "t";
+		actual = geoHashRegion.getSouthEast();
+		assertEquals(expected, actual);
+
+		geoHashRegion = new FirstRingRegion("t");
+
+		expected = "q";
+		actual = geoHashRegion.getSouthEast();
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link FirstRingRegion#getSouthWest()}.
+	 */
+	@Test
+	public void testGetSouthWest()
+	{
+		FirstRingRegion geoHashRegion = new FirstRingRegion("u");
+
+		expected = "e";
+		actual = geoHashRegion.getSouthWest();
+		assertEquals(expected, actual);
+
+		geoHashRegion = new FirstRingRegion("t");
+
+		expected = "k";
+		actual = geoHashRegion.getSouthWest();
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link FirstRingRegion#getWest()}.
+	 */
+	@Test
+	public void testGetWest()
+	{
+		FirstRingRegion geoHashRegion = new FirstRingRegion("u");
+
+		expected = "g";
+		actual = geoHashRegion.getWest();
+		assertEquals(expected, actual);
+
+		geoHashRegion = new FirstRingRegion("t");
+
+		expected = "s";
+		actual = geoHashRegion.getWest();
+		assertEquals(expected, actual);
+	}
+
 }