+++ jgeohash-geoip/src/main/java/de/alpharogroup/jgeohash/geoip/LocationExtensions.java
@@ -139,4 +139,5 @@ public class LocationExtensions
 	{
 		return sortByDistance(startPoint, geohashes, MeasuringUnit.METER);
 	}
+
 }